requirements.txt
fastapi[standard]
passlib[bcrypt]
pyjwt[crypto]
colorlog
aiosmtplib
sqlalchemy
pydantic-settings
aiosqlite
sqladmin
alembic
aiofiles
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client